Update mass, power, and drag
Mass was incorrectly set to the weight of the car with no fuel and no driver; it should include the driver.

Power and drag were set to correspond as closely as possible with the real world telemetry but the wind was set incorrectly; either lilski's Road America is oriented incorrectly (I doubt this) or the wind setting in Content Manager is backwards. Setting the wind to 13km/h S allowed me to get the telemetry to correspond almost exactly to a lap set when KSBM was reporting 35010KT.