Adjust suspension for broken RIM_OFFSET setting
I'm told AC's STRUT suspension type is very buggy, and one of the bugs
in it appears to be that it igores the RIM_OFFSET setting. To
compensate, I set the track 80mm narrower and moved all suspension X
coordinates out by 40mm. To reinforce, the RIM_OFFSET was also set to 0.