Add COSMIC suspension
The implementation is a bit funky. I attempted to compensate for the
displacement of the suspension when loaded in the measurements on the
rear axle and specify an appropriate preload, but on the front I simply
used the full droop measurements. I'll probably go back and make them
consistent. There also seems to be something up with the
spring/damper/ARB values now, as the car doesn't seem quite right though
the geometry is much closer to what it is in reality.